rearranged matrix; breaking ground on templatizing vector with concepts and modules
