+++ experimental/concepts/test_concept.cpp
@@ -1,24 +1,25 @@
 
 #include <print>
-#include <valarray>
 #include "vector_concept.hpp"
 #include "../../vectorspace/vector.hpp"
 
 int main()
 {
   using namespace linalg;
-  static_assert(concepts::vector_c<vector>);
+
   const vector u{1., 0., 0.};
   const vector v{0., 1., 0.};
-  auto lambda = [](concepts::vector_c auto& a, concepts::vector_c auto& b) {
+  auto lambda = [] (concepts::group_element_c_weak auto& a,
+                    concepts::group_element_c_weak auto& b) {
     return a + b;
   };
-  const vector w = lambda(u, v);
+  vector w = lambda(u, v);
   w.print();
   linalg::print(w + w);
-  std::valarray a{1., 0., 1.};
-  std::valarray b{-1., 1., 1.};
+  //std::println("{}", w);
   static_assert(concepts::group_element_c_weak<vector>);
+  static_assert(concepts::group_element_c_weak<double>);
+  static_assert(concepts::field_element_c_weak<double>);
   //std::valarray c = lambda(a, b);
   //std::println(c);
 }
