+++ experimental/template/gcc_headers.hpp
@@ -1,0 +1,16 @@
+
+
+#include "prelude.hpp"
+#include <cstdlib>
+
+#include <cmath>
+#include <initializer_list>
+#include <iostream>
+#include <exception>
+#include <limits>
+#include <list>
+#include <memory>
+#include <numeric>
+#include <string>
+#include <optional>
+#include <vector>
